# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rayzan Mart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin/user)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # My Orders
...
  /me/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get my profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # Loyalty
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        isBlocked:
          type: boolean
        dateOfBirth:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        nid:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        paymentNumber:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, name]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        district:
          type: string
        avatarUrl:
          type: string
        dateOfBirth:
          type: string
        occupation:
          type: string
        nid:
          type: string
        paymentMethod:
          type: string
        paymentNumber:
          type: string
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        userId:
          type: string
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        shippingAddress:
          type: string
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        subtotal:
          type: number
        deliveryCharge:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        deliveryType:
          type: string
        couponCode:
          type: string
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        courier:
          type: string
          nullable: true
        affiliateReferralCode:
          type: string
          nullable: true
        pointsEarned:
          type: integer
        pointsRedeemed:
          type: integer
        pointsDiscountAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNumber, customerName, customerPhone, total, status]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
          nullable: true
        variantId:
          type: string
          nullable: true
        productNameBn:
          type: string
        productNameEn:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        variantAttributes:
          type: object
          nullable: true
      required: [id, orderId, productNameBn, productNameEn, quantity, unitPrice, totalPrice]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, limit, totalPages]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        city:
          type: string
        district:
          type: string
        deliveryType:
          type: string
          enum: [inside_city, outside_city]
        paymentMethod:
          type: string
          enum: [cod, online]
        couponCode:
          type: string
        affiliateReferralCode:
          type: string
        pointsToRedeem:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required: [customerName, customerPhone, shippingAddress, deliveryType, paymentMethod, items]
    OrderItemInput:
      type: object
      properties:
        productId:
          type: string
        variantId:
          type: string
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, returned, cancelled]
        trackingNumber:
          type: string
        courier:
          type: string
        notes:
          type: string
      required: [status]
...