# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyDeliveryType } from "./createOrderBodyDeliveryType";
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  shippingAddress: string;
  city?: string;
  district?: string;
  deliveryType: CreateOrderBodyDeliveryType;
  paymentMethod: CreateOrderBodyPaymentMethod;
  couponCode?: string;
  affiliateReferralCode?: string;
  pointsToRedeem?: number;
  notes?: string;
  items: OrderItemInput[];
}