# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  orderNumber: string;
  userId?: string | null;
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  shippingAddress?: string;
  city?: string | null;
  district?: string | null;
  subtotal?: number;
  deliveryCharge?: number;
  discountAmount?: number;
  total: number;
  status: string;
  paymentMethod?: string;
  deliveryType?: string;
  couponCode?: string | null;
  trackingNumber?: string | null;
  courier?: string | null;
  affiliateReferralCode?: string | null;
  pointsEarned?: number;
  pointsRedeemed?: number;
  pointsDiscountAmount?: number;
  notes?: string | null;
  createdAt?: string;
}